requirements.txt
python-telegram-bot
requests
transliterate
rfc6266_parser